feat(front/admin/products): create restore button, fix state redux isTrashed

+++ packages/frontend/src/pages/admin/products.tsx
@@ -5,6 +5,7 @@ import { useDispatch, useSelector } from "react-redux";
 import { RootState, AppDispatch } from "../../redux/store/index";
 import { useEffect, useState } from "react";
 import * as actions from "../../redux/actions/index";
+import axios from "axios";
 
 const ProductsAdminPage = () => {
     const [currentPage, setCurrentPage] = useState(1);
@@ -38,14 +39,24 @@ const ProductsAdminPage = () => {
         alert(id);
     };
 
-    const handleDelete = (id: number) => {
-        if (window.confirm('¿Está seguro de que desea realizar esta acción?')) {
-            // La acción se ejecutará si el usuario hace clic en "Aceptar"
-            // Coloca aquí la lógica para ejecutar la acción que desea confirmar
-            const found = data.find((e: any) => e.id === id);
-            dispatch(actions.deleteProduct(found));
+    const handleDelete = async(value: any) => {
+        if (value.isTrashed) {
+            await axios.patch(`http://localhost:4000/products/${value.id}`, {
+                ...data,
+                isTrashed: false,
+            });
             dispatch(actions.getProductsPerPage(currentPage))
-          } 
+
+        } else {
+
+            if (window.confirm('¿Está seguro de que desea realizar esta acción?')) {
+                // La acción se ejecutará si el usuario hace clic en "Aceptar"
+                // Coloca aquí la lógica para ejecutar la acción que desea confirmar
+                const found = data.find((e: any) => e.id === value.id);
+                dispatch(actions.deleteProduct(found));
+                dispatch(actions.getProductsPerPage(currentPage))
+            }
+        }
     };
     return (
         <Flex direction="column" alignItems="center">
